+++ src/test/market_and_gear/TestWeapon.java
@@ -19,8 +19,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 public class TestWeapon {
     private final String name = "Sword";
     private final Level minLevel = new Level(10);
-    private final double price = 250;
-    private final double damage = 500;
+    private final int price = 250;
+    private final int damage = 500;
     private final int numHands = 2;
     private final Weapon weapon;
